+++ internal/servers/controller/handlers/accounts/validate_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/hashicorp/boundary/internal/auth/password"
 	pb "github.com/hashicorp/boundary/internal/gen/controller/api/resources/accounts"
 	pbs "github.com/hashicorp/boundary/internal/gen/controller/api/services"
+	"github.com/hashicorp/boundary/internal/intglobals"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/protobuf/types/known/fieldmaskpb"
@@ -154,9 +155,19 @@ func TestValidateUpdateRequest(t *testing.T) {
 		errContains string
 	}{
 		{
-			name: "password to oidc change type",
+			name: "password to oidc change type, old prefix",
 			req: &pbs.UpdateAccountRequest{
-				Id: password.AccountPrefix + "_1234567890",
+				Id: intglobals.OldPasswordAccountPrefix + "_1234567890",
+				Item: &pb.Account{
+					Type: auth.OidcSubtype.String(),
+				},
+			},
+			errContains: fieldError(typeField, "Cannot modify the resource type."),
+		},
+		{
+			name: "password to oidc change type, new prefix",
+			req: &pbs.UpdateAccountRequest{
+				Id: intglobals.NewPasswordAccountPrefix + "_1234567890",
 				Item: &pb.Account{
 					Type: auth.OidcSubtype.String(),
 				},
@@ -174,9 +185,21 @@ func TestValidateUpdateRequest(t *testing.T) {
 			errContains: fieldError(typeField, "Cannot modify the resource type."),
 		},
 		{
-			name: "password bad attributes",
+			name: "password bad attributes old prefix",
 			req: &pbs.UpdateAccountRequest{
-				Id: password.AccountPrefix + "_1234567890",
+				Id: intglobals.OldPasswordAccountPrefix + "_1234567890",
+				Item: &pb.Account{
+					Attributes: &structpb.Struct{Fields: map[string]*structpb.Value{
+						"test": structpb.NewStringValue("something"),
+					}},
+				},
+			},
+			errContains: fieldError(attributesField, "Attribute fields do not match the expected format."),
+		},
+		{
+			name: "password bad attributes new prefix",
+			req: &pbs.UpdateAccountRequest{
+				Id: intglobals.NewPasswordAccountPrefix + "_1234567890",
 				Item: &pb.Account{
 					Attributes: &structpb.Struct{Fields: map[string]*structpb.Value{
 						"test": structpb.NewStringValue("something"),